move logging functions to core.py; improve docstrings
