+++ src/rtc/core.py
@@ -1,0 +1,100 @@
+#!/usr/bin/env python
+
+import sys
+import logging
+
+logger = logging.getLogger('rtc_s1')
+
+class Logger(object):
+    """
+    Class to redirect stdout and stderr to the logger
+    """
+    def __init__(self, logger, level, prefix=''):
+       """
+       Class constructor
+       """
+       self.logger = logger
+       self.level = level
+       self.prefix = prefix
+       self.buffer = ''
+
+    def write(self, message):
+
+        # Add message to the buffer until "\n" is found
+        if '\n' not in message:
+            self.buffer += message
+            return
+
+        message = self.buffer + message
+
+        # check if there is any character after the last \n
+        # if so, move it to the buffer
+        message_list = message.split('\n')
+        if not message.endswith('\n'):
+            self.buffer = message_list[-1]
+            message_list = message_list[:-1]
+        else:
+            self.buffer = ''
+
+        # print all characters before the last \n
+        for line in message_list:
+            if not line:
+                continue
+            self.logger.log(self.level, self.prefix + line)
+
+    def flush(self):
+        self.logger.log(self.level, self.buffer)
+        self.buffer = ''
+
+
+def create_logger(log_file, full_log_formatting=None):
+    """Create logger object for a log file
+
+       Parameters
+       ----------
+       log_file: str
+              Log file
+       full_log_formatting : bool
+              Flag to enable full formatting of logged messages
+
+       Returns
+       -------
+       logger : logging.Logger
+              Logger object
+    """
+    # create logger
+    logger.setLevel(logging.DEBUG)
+
+    # create console handler and set level to debug
+    ch = logging.StreamHandler()
+    ch.setLevel(logging.DEBUG)
+
+    # create formatter
+    # configure full log format, if enabled
+    if full_log_formatting:
+        msgfmt = ('%(asctime)s.%(msecs)03d, %(levelname)s, DSWx-HLS, '
+                  '%(module)s, 999999, %(pathname)s:%(lineno)d, "%(message)s"')
+
+        formatter = logging.Formatter(msgfmt, "%Y-%m-%d %H:%M:%S")
+    else:
+        formatter = logging.Formatter('%(message)s')
+
+    # add formatter to ch
+    ch.setFormatter(formatter)
+
+    # add ch to logger
+    logger.addHandler(ch)
+
+    if log_file:
+        file_handler = logging.FileHandler(log_file)
+
+        file_handler.setFormatter(formatter)
+
+        # add file handler to logger
+        logger.addHandler(file_handler)
+
+    sys.stdout = Logger(logger, logging.INFO)
+    sys.stderr = Logger(logger, logging.ERROR, prefix='[StdErr] ')
+
+    return logger
+
